backup/requirements.txt
python-dotenv>=1.0.0
croniter>=2.0.0
requests>=2.31.0
